Maintenance

+++ Sirius/src/Sirius/Events/MouseEvent.h
@@ -7,11 +7,11 @@ namespace Sirius {
 	class MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(double x, double y)
+		MouseMovedEvent(float x, float y)
 			: m_MouseX(x), m_MouseY(y) {}
 
-		inline double GetX() const { return m_MouseX; }
-		inline double GetY() const { return m_MouseY; }
+		inline float GetX() const { return m_MouseX; }
+		inline float GetY() const { return m_MouseY; }
 
 		std::string ToString() const override
 		{
@@ -24,17 +24,17 @@ namespace Sirius {
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 
 	private:
-		double m_MouseX, m_MouseY;
+		float m_MouseX, m_MouseY;
 	};
 
 	class MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(double xOffset, double yOffset)
+		MouseScrolledEvent(float xOffset, float yOffset)
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		inline double GetXOffset() const { return m_XOffset; }
-		inline double GetYOffset() const { return m_YOffset; }
+		inline float GetXOffset() const { return m_XOffset; }
+		inline float GetYOffset() const { return m_YOffset; }
 
 		std::string ToString() const override
 		{
@@ -47,7 +47,7 @@ namespace Sirius {
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 
 	private:
-		double m_XOffset, m_YOffset;
+		float m_XOffset, m_YOffset;
 	};
 
 	class MouseButtonEvent : public Event
